+++ lib/src/main/java/com/aeontronix/enhancedmule/tools/exchange/AssetType.java
@@ -1,0 +1,9 @@
+/*
+ * Copyright (c) Aeontronix 2020
+ */
+
+package com.aeontronix.enhancedmule.tools.exchange;
+
+public enum AssetType {
+    CUSTOM, CONNECTOR, APPLICATION, EXAMPLE, TEMPLATE, HTTP_SPEC, RAML_SPEC, OAS2_SPEC, OAS3_SPEC
+}